Function libsignal_protocol_sys::signal_protocol_key_helper_generate_pre_keys

pub unsafe extern "C" fn signal_protocol_key_helper_generate_pre_keys(
    head: *mut *mut signal_protocol_key_helper_pre_key_list_node, 
    start: c_uint, 
    count: c_uint, 
    global_context: *mut signal_context
) -> c_int

Generate a list of PreKeys. Clients should do this at install time, and subsequently any time the list of PreKeys stored on the server runs low.

Pre key IDs are shorts, so they will eventually be repeated. Clients should store pre keys in a circular buffer, so that they are repeated as infrequently as possible.

When finished with this list, the caller should free it by calling signal_protocol_key_helper_key_list_free().

@param head pointer to the head of the key list @param start the starting pre key ID, inclusive. @param count the number of pre keys to generate. @return 0 on success, or negative on failure