Function libsignal_protocol_sys::signal_protocol_key_helper_generate_signed_pre_key

pub unsafe extern "C" fn signal_protocol_key_helper_generate_signed_pre_key(
    signed_pre_key: *mut *mut session_signed_pre_key, 
    identity_key_pair: *const ratchet_identity_key_pair, 
    signed_pre_key_id: u32, 
    timestamp: u64, 
    global_context: *mut signal_context
) -> c_int

Generate a signed pre key

@param signed_pre_key set to the generated pre key @param identity_key_pair the local client's identity key pair. @param signed_pre_key_id the pre key ID to assign the generated signed pre key @param timestamp the current time in milliseconds since the UNIX epoch

@return 0 on success, or negative on failure