Module libsignal_protocol_sys::protobuf

Structs

ProtobufCAllocator
Structure for defining a custom memory allocator.

ProtobufCBuffer
Structure for defining a virtual append-only buffer. Used by protobuf_c_message_pack_to_buffer() to abstract the consumption of serialized bytes.

ProtobufCEnumDescriptor
Describes an enumeration as a whole, with all of its values.

ProtobufCEnumValue
Represents a single value of an enumeration.

ProtobufCEnumValueIndex
Used by ProtobufCEnumDescriptor to look up enum values.

ProtobufCFieldDescriptor
Describes a single field in a message.

ProtobufCIntRange
Helper structure for optimizing int => index lookups in the case where the keys are mostly consecutive values, as they presumably are for enums and fields.

ProtobufCMessage
An instance of a message.

ProtobufCMessageDescriptor
Describes a message.

ProtobufCMessageUnknownField
An unknown message field.

ProtobufCMethodDescriptor
Method descriptor.

ProtobufCService
Service.

ProtobufCServiceDescriptor
Service descriptor.

Constants

ProtobufCLabel_PROTOBUF_C_LABEL_OPTIONAL
A well-formed message can have zero or one of this field (but not more than one).

ProtobufCLabel_PROTOBUF_C_LABEL_REPEATED
This field can be repeated any number of times (including zero) in a well-formed message. The order of the repeated values will be preserved.

ProtobufCLabel_PROTOBUF_C_LABEL_REQUIRED
A well-formed message must have exactly one of this field.

ProtobufCType_PROTOBUF_C_TYPE_BOOL
< boolean

ProtobufCType_PROTOBUF_C_TYPE_BYTES
< arbitrary byte sequence

ProtobufCType_PROTOBUF_C_TYPE_DOUBLE
< double

ProtobufCType_PROTOBUF_C_TYPE_ENUM
< enumerated type

ProtobufCType_PROTOBUF_C_TYPE_FIXED32
< unsigned int32 (4 bytes)

ProtobufCType_PROTOBUF_C_TYPE_FIXED64
< unsigned int64 (8 bytes)

ProtobufCType_PROTOBUF_C_TYPE_FLOAT
< float

ProtobufCType_PROTOBUF_C_TYPE_INT32
< int32

ProtobufCType_PROTOBUF_C_TYPE_INT64
< int64

ProtobufCType_PROTOBUF_C_TYPE_MESSAGE
< nested message

ProtobufCType_PROTOBUF_C_TYPE_SFIXED32
< signed int32 (4 bytes)

ProtobufCType_PROTOBUF_C_TYPE_SFIXED64
< signed int64 (8 bytes)

ProtobufCType_PROTOBUF_C_TYPE_SINT32
< signed int32

ProtobufCType_PROTOBUF_C_TYPE_SINT64
< signed int64

ProtobufCType_PROTOBUF_C_TYPE_STRING
< UTF-8 or ASCII string

ProtobufCType_PROTOBUF_C_TYPE_UINT32
< unsigned int32

ProtobufCType_PROTOBUF_C_TYPE_UINT64
< unsigned int64

ProtobufCWireType_PROTOBUF_C_WIRE_TYPE_32BIT

ProtobufCWireType_PROTOBUF_C_WIRE_TYPE_64BIT

ProtobufCWireType_PROTOBUF_C_WIRE_TYPE_LENGTH_PREFIXED

ProtobufCWireType_PROTOBUF_C_WIRE_TYPE_VARINT

Functions

protobuf_c_buffer_simple_append⚠
The append method for ProtobufCBufferSimple.

protobuf_c_enum_descriptor_get_value⚠
Look up a ProtobufCEnumValue from a ProtobufCEnumDescriptor by numeric value.

protobuf_c_enum_descriptor_get_value_by_name⚠
Look up a ProtobufCEnumValue from a ProtobufCEnumDescriptor by name.

protobuf_c_message_check⚠
Check the validity of a message object.

protobuf_c_message_descriptor_get_field⚠
Look up a ProtobufCFieldDescriptor from a ProtobufCMessageDescriptor by the tag value of the field.

protobuf_c_message_descriptor_get_field_by_name⚠
Look up a ProtobufCFieldDescriptor from a ProtobufCMessageDescriptor by the name of the field.

protobuf_c_message_free_unpacked⚠
Free an unpacked message object.

protobuf_c_message_get_packed_size⚠
Determine the number of bytes required to store the serialised message.

protobuf_c_message_init⚠
Initialise a message object from a message descriptor.

protobuf_c_message_pack⚠
Serialise a message from its in-memory representation.

protobuf_c_message_pack_to_buffer⚠
Serialise a message from its in-memory representation to a virtual buffer.

protobuf_c_message_unpack⚠
Unpack a serialised message into an in-memory representation.

protobuf_c_service_descriptor_get_method_by_name⚠
Look up a ProtobufCMethodDescriptor by name.

protobuf_c_service_destroy⚠
Free a service.

protobuf_c_service_generated_init⚠

protobuf_c_service_invoke_internal⚠

protobuf_c_version⚠
Get the version of the protobuf-c library. Note that this is the version of the library linked against, not the version of the headers compiled against.

protobuf_c_version_number⚠
Get the version of the protobuf-c library. Note that this is the version of the library linked against, not the version of the headers compiled against.

Type Definitions

ProtobufCClosure

ProtobufCLabel
Message field rules.

ProtobufCMessageInit

ProtobufCServiceDestroy

ProtobufCType
Field value types.

ProtobufCWireType
Field wire types.

protobuf_c_boolean
Boolean type.