Struct libsignal_protocol_sys::protobuf::ProtobufCFieldDescriptor

#[repr(C)]

pub struct ProtobufCFieldDescriptor {
    pub name: *const c_char,
    pub id: u32,
    pub label: ProtobufCLabel,
    pub type_: ProtobufCType,
    pub quantifier_offset: c_uint,
    pub offset: c_uint,
    pub descriptor: *const c_void,
    pub default_value: *const c_void,
    pub flags: u32,
    pub reserved_flags: c_uint,
    pub reserved2: *mut c_void,
    pub reserved3: *mut c_void,
}

Describes a single field in a message.

Fields
name: *const c_char

Name of the field as given in the .proto file.

id: u32

Tag value of the field as given in the .proto file.

label: ProtobufCLabel

Whether the field is REQUIRED, OPTIONAL, or REPEATED.

type_: ProtobufCType

The type of the field.

quantifier_offset: c_uint

The offset in bytes of the message's C structure's quantifier field (the has_MEMBER field for optional members or the n_MEMBER field for repeated members or the case enum for oneofs).

offset: c_uint

The offset in bytes into the message's C structure for the member itself.

descriptor: *const c_void

A type-specific descriptor.

If type is PROTOBUF_C_TYPE_ENUM, then descriptor points to the corresponding ProtobufCEnumDescriptor.

If type is PROTOBUF_C_TYPE_MESSAGE, then descriptor points to the corresponding ProtobufCMessageDescriptor.

Otherwise this field is NULL.

default_value: *const c_void

The default value for this field, if defined. May be NULL.

flags: u32

A flag word. Zero or more of the bits defined in the ProtobufCFieldFlag enum may be set.

reserved_flags: c_uint

Reserved for future use.

reserved2: *mut c_void

Reserved for future use.

reserved3: *mut c_void

Reserved for future use.